+++ app/Http/Controllers/Admin/PartnerController.php
@@ -4,8 +4,8 @@ namespace app\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Partner;
-use App\PartnerCategory;
+use App\Models\Partner;
+use App\Models\PartnerCategory;
 use Validator;
 
 class PartnerController extends Controller
