+++ app/Models/Partner.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace app\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -9,6 +9,4 @@ class Partner extends Model
     protected $fillable = [
         'name', 'introduction', 'logo_url', 'connect',
     ];
-
-    
 }
