--- app/PartnerCategory.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class PartnerCategory extends Model
-{
-     protected $table = 'partner_categories';
-	 protected $fillable = ['name'];
-}
